requirements/dev.txt
# path: requirements/dev.txt
-r base.txt
pytest>=8.3,<9.0
pytest-django>=4.9,<5.0
pytest-cov>=6.0,<7.0
factory_boy>=3.3,<4.0
ruff>=0.9,<1.0
black>=24.10,<25.0
scikit-learn>=1.5,<2.0
